Move USART receiver code to dedicated module
